+++ frontend/src/routes.ts
@@ -6,7 +6,7 @@ import {
 export default [
   route("", "./routes/header.tsx", [
     route("login", "./routes/login.tsx"),
-    route("createUser", "./routes/createUser.tsx")
+    route("createUser", "./routes/createUserMenu.tsx")
   ]),
